requirements.txt
Django==1.9.1
behave-django==0.3.0
splinter==0.7.3
